src/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
databricks-sdk>=0.72.0
requests>=2.31.0
